Shared: Move util to shared
